BLE Refactoring - Validated
- Moved experimental (unused) BLE code
