+++ herald/src/main/java/com/vmware/herald/sensor/ble/filter/BLEAdvertManufacturerData.java
@@ -2,7 +2,7 @@
 //  SPDX-License-Identifier: MIT
 //
 
-package com.vmware.herald.sensor.ble;
+package com.vmware.herald.sensor.ble.filter;
 
 public class BLEAdvertManufacturerData {
     public int manufacturer;
