+++ herald/src/main/java/com/vmware/herald/sensor/ble/filter/BLEAdvertSegment.java
@@ -2,7 +2,7 @@
 //  SPDX-License-Identifier: MIT
 //
 
-package com.vmware.herald.sensor.ble;
+package com.vmware.herald.sensor.ble.filter;
 
 public class BLEAdvertSegment {
     public BLEAdvertSegmentType type;
